+++ src/pages/TechnicalPage/TechnicalPage.tsx
@@ -48,6 +48,7 @@ export default function TechnicalPage() {
 
   const getProposal = () => application.content2 as Proposal;
   const setProposal = (proposal: Proposal) => updateAppContent2(proposal);
+  const authClient = useAxiosAuthClient();
 
   const getProposalState = () => application.content1 as number[];
   const setTheProposalState = (value: number) => {
@@ -79,7 +80,7 @@ export default function TechnicalPage() {
 
     try {
       const proposal = getProposal();
-      const signedUrl = await GetPresignedUploadUrl(`${proposal.id}-technical.pdf`);
+      const signedUrl = await GetPresignedUploadUrl(authClient, `${proposal.id}-technical.pdf`);
 
       if (typeof signedUrl != 'string') new Error('Not able to Get Technical PDF Upload URL');
 
@@ -110,7 +111,7 @@ export default function TechnicalPage() {
     try {
       const proposal = getProposal();
       const selectedFile = `${proposal.id}-` + t('pdfDownload.technical.label') + t('fileType.pdf');
-      const signedUrl = await GetPresignedDownloadUrl(selectedFile);
+      const signedUrl = await GetPresignedDownloadUrl(authClient, selectedFile);
 
       if (signedUrl === t('pdfDownload.sampleData') || proposal.technicalPDF != null) {
         window.open(signedUrl, '_blank');
@@ -123,7 +124,7 @@ export default function TechnicalPage() {
   const deletePdfUsingSignedUrl = async () => {
     try {
       const proposal = getProposal();
-      const signedUrl = await GetPresignedDeleteUrl(`${proposal.id}-technical.pdf`);
+      const signedUrl = await GetPresignedDeleteUrl(authClient, `${proposal.id}-technical.pdf`);
 
       if (typeof signedUrl != 'string') new Error('Not able to Get Technical PDF Upload URL');
 
@@ -155,7 +156,7 @@ export default function TechnicalPage() {
     try {
       const proposal = getProposal();
       const selectedFile = `${proposal.id}-` + t('pdfDownload.technical.label') + t('fileType.pdf');
-      const signedUrl = await GetPresignedDownloadUrl(selectedFile);
+      const signedUrl = await GetPresignedDownloadUrl(authClient, selectedFile);
 
       if (signedUrl === t('pdfDownload.sampleData') || proposal.technicalPDF != null) {
         setCurrentFile(signedUrl);