+++ src/services/axios/postSendEmailInvite/mockEmailInvite.tsx
@@ -1,0 +1,6 @@
+import { EmailInviteBackend } from '@utils/types/emailInvite.tsx';
+
+export const MockEmailInvite: EmailInviteBackend = {
+  email: 'chloe.gallacher@cgi.com',
+  prsl_id: 'prsl-t0001-20250806-00002'
+};