+++ src/services/axios/postSendEmailInvite/postSendEmailInvite.tsx
@@ -1,4 +1,4 @@
-import { OSO_SERVICES_PROPOSAL_PATH, SKA_OSO_SERVICES_URL } from '../../../utils/constants';
+import { OSO_SERVICES_PROPOSAL_PATH, SKA_OSO_SERVICES_URL } from '@utils/constants.ts';
 import useAxiosAuthClient from '../axiosAuthClient/axiosAuthClient';
 import { EmailInviteBackend } from '@/utils/types/emailInvite.tsx';
 
@@ -7,7 +7,7 @@ async function PostSendEmailInvite(
   email: EmailInviteBackend
 ) {
   try {
-    const URL_PATH = `${OSO_SERVICES_PROPOSAL_PATH}/send-email`;
+    const URL_PATH = `${OSO_SERVICES_PROPOSAL_PATH}/send-email/`;
 
     const result = await authAxiosClient.post(`${SKA_OSO_SERVICES_URL}${URL_PATH}`, email);
     return typeof result === 'undefined' ? 'error.API_UNKNOWN_ERROR' : result.data;